Build and update code for repository of User

+++ Code/company/USR/User/repository/VSoft.Company.USR.User.Repository.Efc.Provider/Services/EfcUserRepository.cs
@@ -6,7 +6,7 @@ using VSoft.Company.USR.User.Repository.Efc.Services;
 
 namespace VSoft.Company.USR.User.Repository.Efc.Provider.Services;
 
-public class EfcUserRepository : EFcRepositoryEntityMgmtId<UserDbContext, MUserEntity, long>, IUserRepositoryEfc
+public class EfcUserRepository : EFcRepositoryEntityMgmtId<UserDbContext, MUserEntity, int>, IUserRepositoryEfc
 {
 
     public EfcUserRepository(UserDbContext dbContext) : base(dbContext, dbContext.Items)
@@ -14,7 +14,7 @@ public class EfcUserRepository : EFcRepositoryEntityMgmtId<UserDbContext, MUserE
 
     }
 
-    public string? GetFullName(long? id)
+    public string? GetFullName(int? id)
     {
         if (DbContext == null) throw new Exception("Context is null");
         if (Entities == null) throw new Exception("Entities is null");
@@ -22,7 +22,7 @@ public class EfcUserRepository : EFcRepositoryEntityMgmtId<UserDbContext, MUserE
         return Entities.Where(x => x.Id == id).Select(x => x.Name ?? string.Empty).FirstOrDefault();
     }
 
-    public Task<string?> GetFullNameAsync(long? id)
+    public Task<string?> GetFullNameAsync(int? id)
     {
         if (DbContext == null) throw new Exception("Context is null");
         if (Entities == null) throw new Exception("Entities is null");
