+++ Code/company/USR/User/repository/VSoft.Company.USR.User.Repository.Efc/Services/IUserRepositoryEfc.cs
@@ -5,7 +5,7 @@ using VSoft.Company.USR.User.Repository.Services;
 
 namespace VSoft.Company.USR.User.Repository.Efc.Services;
 
-public interface IUserRepositoryEfc : IUserRepository, IEfcRepositoryEntityMgmtId<UserDbContext, MUserEntity, long>
+public interface IUserRepositoryEfc : IUserRepository, IEfcRepositoryEntityMgmtId<UserDbContext, MUserEntity, int>
 {
    
 }
